# Speedup expand when meets large image