--- a/mmdet/datasets/pipelines/transforms.py
+++ b/mmdet/datasets/pipelines/transforms.py
@@ -854,7 +854,8 @@ class Expand(object):
 
         h, w, c = img.shape
         ratio = random.uniform(self.min_ratio, self.max_ratio)
-        if self.mean[0] == self.mean[1] and self.mean[1] == self.mean[2]:
+        # speedup expand when meets large image
+        if np.all(self.mean == self.mean[0]):
             expand_img = np.empty((int(h * ratio), int(w * ratio), c),
                                   img.dtype)
             expand_img.fill(self.mean[0])