ref: https://github.com/numpy/numpy/issues/13001
```python
%%timeit
a = np.full((4000, 2000, 3), [0.0, 0.0, 0.0], np.float32)
76.9 ms ± 346 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
%%timeit
a = np.full((4000, 2000, 3), [0, 0, 0], np.float32)
71.4 ms ± 516 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
%%timeit
a = np.full((4000, 2000, 3), 0.0, np.float32)
47.2 ms ± 1.21 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
%%timeit
a = np.full((4000, 2000, 3), 0, np.float32)
47 ms ± 139 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
%%timeit
a = np.empty((4000, 2000, 3), np.float32)
a.fill(0.0)
46.3 ms ± 278 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
%%timeit
a = np.empty((4000, 2000, 3), np.float32)
a.fill(0)
46.3 ms ± 184 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```